// frontend/src/types/protobufTypeUrls.ts
export enum ProtobufTypeUrls {
  AUTHENTICATION_METADATA = 'type.googleapis.com/buildbarn.auth.AuthenticationMetadata',
  REQUEST_METADATA = 'type.googleapis.com/build.bazel.remote.execution.v2.RequestMetadata',
  POSIX_RESOURCE_USAGE = 'type.googleapis.com/buildbarn.resourceusage.POSIXResourceUsage',
  FILE_POOL_RESOURCE_USAGE = 'type.googleapis.com/buildbarn.resourceusage.FilePoolResourceUsage',
  INPUT_ROOT_RESOURCE_USAGE = 'type.googleapis.com/buildbarn.resourceusage.InputRootResourceUsage',
  MONETARY_RESOURCE_USAGE = 'type.googleapis.com/buildbarn.resourceusage.MonetaryResourceUsage',
}